requirements.txt
streamlit
gtts
moviepy
Pillow
